requirements.txt
openai-whisper
cohere
pyttsx3
SpeechRecognition
